Fix race conditions in Capture/CaptureWithCGo
When run in parallel, the current code can panic
due to "unlock of unlocked mutex".
Specifically, what would happen is:

- Routine1 swaps file descriptors
- Routine2 then swaps FDs again
- Routine1 finishes and tries to restore FDs
- Routine1 errors restoring FDs in the defer,
  causing one of the 'if e != nil, unlock' branches
  to be hit.
- The unconditional 'unlock' call then runs,
  and re-unlocks the mutex, panicing.

Because there were no tests with concurrency,
this bug wasn't detected (it is trivial to reproduce
by adding parallelization, as the test indicates).

While trying to debug and fix this bug (avoiding multiple unlocks),
I realized that the cause of hitting the branch (failure restoring FDs)
indicates a logical problem- because the FDs are global to the process,
it would be very easy to lose/interleave input:

- Routine1 swaps FDs (stdout/err are routine1's copies)
- Routine2 swaps FDs (stdout/err are routine2's copies)
- Routine1's callback runs and writes to routine2's fd copies

To avoid this, I simplified everything and take a mutex on the whole
function (multiple defers run in LIFO).
The downside here is obviously that we have to run the callback
inside of the mutex; however, I am not sure what the option could be.
If code (C or Go) writes to stdout/os.Stdout or err,
it logically must run in isolation.
Avoiding the function-side lock allows concurrent callbacks
but with incorrect results.